Refactor routing and controllers.
