+++ app-modules/portal/src/Http/Controllers/KnowledgeManagement/KnowledgeManagementPortalArticleController.php
@@ -34,26 +34,31 @@
 </COPYRIGHT>
 */
 
-use Illuminate\Support\Facades\Route;
-use AdvisingApp\Portal\Livewire\RenderKnowledgeManagementPortal;
-use AdvisingApp\Portal\Http\Middleware\EnsureKnowledgeManagementPortalIsEnabled;
-use AdvisingApp\Portal\Http\Middleware\EnsureKnowledgeManagementPortalIsEmbeddableAndAuthorized;
+namespace AdvisingApp\Portal\Http\Controllers\KnowledgeManagement;
 
-Route::prefix('portals')
-    ->name('portals.')
-    ->middleware([
-        'web',
-    ])
-    ->group(function () {
-        Route::middleware([
-            EnsureKnowledgeManagementPortalIsEnabled::class,
-            EnsureKnowledgeManagementPortalIsEmbeddableAndAuthorized::class,
-        ])->group(function () {
-            Route::get('/knowledge-management/{category?}/{article?}', RenderKnowledgeManagementPortal::class)
-                ->where([
-                    'category' => '[a-zA-Z0-9-]+',
-                    'article' => '[a-zA-Z0-9-]+',
-                ])
-                ->name('knowledge-management.show');
-        });
-    });
+use Illuminate\Http\JsonResponse;
+use App\Http\Controllers\Controller;
+use AdvisingApp\KnowledgeBase\Models\KnowledgeBaseItem;
+use AdvisingApp\KnowledgeBase\Models\KnowledgeBaseCategory;
+use AdvisingApp\Portal\DataTransferObjects\KnowledgeBaseItemData;
+use AdvisingApp\Portal\DataTransferObjects\KnowledgeBaseCategoryData;
+
+class KnowledgeManagementPortalArticleController extends Controller
+{
+    public function show(KnowledgeBaseCategory $category, KnowledgeBaseItem $article): JsonResponse
+    {
+        return response()->json([
+            'category' => KnowledgeBaseCategoryData::from([
+                'id' => $category->getKey(),
+                'name' => $category->name,
+                'description' => $category->description,
+            ]),
+            'article' => KnowledgeBaseItemData::from([
+                'id' => $article->getKey(),
+                'name' => $article->title,
+                'lastUpdated' => $article->updated_at->format('M d Y, h:m a'),
+                'content' => tiptap_converter()->asHTML($article->article_details),
+            ]),
+        ]);
+    }
+}
